LiteFlow-AI/backend/pyproject.toml
[project]
name = "liteflow-ai-backend"
version = "0.1.0"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0.post1",
    "langchain==1.2.0",
    "langgraph==1.0.5",
    "chromadb==0.4.19",
    "ollama==0.2.0",
    "openai==1.3.5",
    "anthropic>=0.3.0,<0.4.0",
    "unstructured>=0.10.0,<0.11.0",
    "pdfplumber==0.10.2",
    "pydantic>=2.7.4,<3.0.0",
    "pydantic-settings==2.0.3",
    "tenacity==8.2.3",
    "httpx==0.27.0",
    "structlog==23.2.0",
    "prometheus-client==0.19.0",
    "opentelemetry-api==1.22.0",
    "opentelemetry-sdk==1.22.0",
    "opentelemetry-exporter-prometheus==0.41b0",
    "opentelemetry-exporter-jaeger==1.18.0",
    "opentelemetry-instrumentation-fastapi==0.43b0",
    "python-dotenv==1.0.0",
    "numpy<2.0.0",
    "markdown==3.5.1",
    "python-multipart==0.0.21"
]

[build-system]
requires = ["uv"]
build-backend = "uv.build"
